picom: remove deprecated refresh-rate, change window focus method
This seems to fix issue of active window not being 'active' (neither
polybar nor picom recognizing the active window changed despite it
having focus and receiving keyboard input properly), at least for
picom.

Polybar still doesn't get the updated window title in the xwindow
module at times (e.g. switching between emacs frames seems to be the
main (only?) case when this happens). That module does use
'_NET_ACTIVE_WINDOW' still, so that seems to be part of the puzzle.

Stumpwm's events.lisp is where to look to see why that isn't being
updated perhaps.

Update: removing the emacs frame that wasn't registered and making a
new frame seems to fix it. Or at least in a quick check. Might be
something particular to emacs frames?